Wrap scripts into a CLI tool
