+++ Rakefile
@@ -1,9 +1,10 @@
-require 'rake/testtask'
+require "bundler/gem_tasks"
+require "rake/testtask"
 
 Rake::TestTask.new(:test) do |t|
-  t.libs << 'test'
-  t.libs << 'lib/deployer'
-  t.test_files = FileList['test/**/*_test.rb']
+  t.libs << "test"
+  t.libs << "lib/deployer"
+  t.test_files = FileList["test/**/*_test.rb"]
 end
 
 task default: :test
