+++ action.yml
@@ -1,0 +1,11 @@
+name: "ServerlessTools"
+description: "Manage Serverless functions."
+inputs:
+  command:
+    description: "which command to run via the serverless-tools cli"
+    required: true
+runs:
+  using: "docker"
+  image: "Dockerfile"
+  args:
+    - ${{ inputs.command }}
