--- actions/comment/action.yml
@@ -1,12 +1,0 @@
-name: "Comment"
-description: "Comments which functions have been deployed on a Github Issue"
-inputs:
-  functions:
-    description: "Hash of functions"
-    required: true
-runs:
-  using: "docker"
-  image: "../../Dockerfile"
-  entrypoint: "lib/comment.rb"
-  args:
-    - ${{ inputs.functions }}
