+++ lib/serverless-tools.rb
@@ -1,0 +1,4 @@
+require "serverless-tools/version"
+require "serverless-tools/cli"
+
+module ServerlessTools; end
