+++ lib/serverless-tools/version.rb
@@ -1,0 +1,3 @@
+module ServerlessTools
+  VERSION = "0.0.2"
+end
