+++ test/deployer/deployer_test.rb
@@ -1,39 +1,40 @@
-require 'minitest/autorun'
-require 'mocha/minitest'
+require "minitest/autorun"
+require "mocha/minitest"
 
-require 'aws-sdk-lambda'
-require 'aws-sdk-s3'
+require "aws-sdk-lambda"
+require "aws-sdk-s3"
 
-require 'deployer/deployer'
-require 'aws_lambda_config'
+require "serverless-tools/deployer/deployer"
+require "serverless-tools/deployer/aws_lambda_config"
 
-describe 'Deployer' do
+describe "Deployer" do
   let(:s3_object_does_not_exist) do
     Aws::S3::Client.new(stub_responses: {
-                          head_object: { status_code: 404, headers: {}, body: '' },
-                          put_object: { status_code: 200, headers: {}, body: '' }
+                          head_object: { status_code: 404, headers: {}, body: "" },
+                          put_object: { status_code: 200, headers: {}, body: "" }
                         })
   end
   let(:s3_object_does_exist) do
     Aws::S3::Client.new(stub_responses: {
-                          head_object: { status_code: 200, headers: {}, body: '' }
+                          head_object: { status_code: 200, headers: {}, body: "" }
                         })
   end
 
   let(:lambda_client) { Aws::Lambda::Client.new(stub_responses: true) }
 
   let(:config) do
-    Deployer::AwsLambdaConfig.new(
-      filename: 'test/fixtures/functions.yml',
-      function_name: 'example_function_one_v1'
+    ServerlessTools::Deployer::AwsLambdaConfig.new(
+      filename: "test/fixtures/functions.yml",
+      function_name: "example_function_one_v1"
     )
   end
 
-  describe '#push' do
+  describe "#push" do
     it "uploads file to S3 object when object doesn't exist" do
-      config.expects(:git_sha).returns('1234567890')
+      config.expects(:git_sha).returns("1234567890")
 
-      uploader = Deployer::Deployer.new(config, s3_client: s3_object_does_not_exist, lambda_client: lambda_client)
+      uploader = ServerlessTools::Deployer::Deployer.new(config, s3_client: s3_object_does_not_exist,
+lambda_client: lambda_client)
       uploader.build
       uploader.push
       File.delete(config.local_filename) if File.exist?(config.local_filename)
@@ -47,14 +48,15 @@ describe 'Deployer' do
       #  with correct params. While this is not strictly required (it tests an
       # internal detail of the class) it is potentially instructive so is left
       # in this first test, but not the subsequent ones
-      head_object_requests = s3_object_does_not_exist.api_requests.select { |x| x[:operation_name] == :head_object }
+      head_object_requests = s3_object_does_not_exist.api_requests.select { |x|
+ x[:operation_name] == :head_object }
       assert_equal(1, head_object_requests.length)
       assert_equal(
-        'freeagent-lambda-example-scripts',
+        "freeagent-lambda-example-scripts",
         head_object_requests.first[:params][:bucket]
       )
       assert_equal(
-        'serverless-tools/deployments/1234567890/example_function_one_v1/function.zip',
+        "serverless-tools/deployments/1234567890/example_function_one_v1/function.zip",
         head_object_requests.first[:params][:key]
       )
 
@@ -63,19 +65,20 @@ describe 'Deployer' do
       put_object_requests = s3_object_does_not_exist.api_requests.select { |x| x[:operation_name] == :put_object }
       assert_equal(1, put_object_requests.length)
       assert_equal(
-        'freeagent-lambda-example-scripts',
+        "freeagent-lambda-example-scripts",
         put_object_requests.first[:params][:bucket]
       )
       assert_equal(
-        'serverless-tools/deployments/1234567890/example_function_one_v1/function.zip',
+        "serverless-tools/deployments/1234567890/example_function_one_v1/function.zip",
         put_object_requests.first[:params][:key]
       )
     end
 
-    it 'does not upload file to S3 object when object exists' do
-      config.expects(:git_sha).returns('1234567890')
+    it "does not upload file to S3 object when object exists" do
+      config.expects(:git_sha).returns("1234567890")
 
-      uploader = Deployer::Deployer.new(config, s3_client: s3_object_does_exist, lambda_client: lambda_client)
+      uploader = ServerlessTools::Deployer::Deployer.new(config, s3_client: s3_object_does_exist,
+lambda_client: lambda_client)
       uploader.push
 
       # see comments above for checking the requests
@@ -86,27 +89,28 @@ describe 'Deployer' do
     end
   end
 
-  describe '#update' do
-    it 'updates lambda code if code object exists' do
-      config.expects(:git_sha).returns('1234567890')
+  describe "#update" do
+    it "updates lambda code if code object exists" do
+      config.expects(:git_sha).returns("1234567890")
       lambda_client
         .expects(:update_function_code)
         .with(
           has_entries(
-            s3_bucket: 'freeagent-lambda-example-scripts',
-            s3_key: 'serverless-tools/deployments/1234567890/example_function_one_v1/function.zip'
+            s3_bucket: "freeagent-lambda-example-scripts",
+            s3_key: "serverless-tools/deployments/1234567890/example_function_one_v1/function.zip"
           )
-        ).returns({ function_name: 'test-function', last_update_status: 'Successful' })
+        ).returns({ function_name: "test-function", last_update_status: "Successful" })
 
-      uploader = Deployer::Deployer.new(config, s3_client: s3_object_does_exist, lambda_client: lambda_client)
+      uploader = ServerlessTools::Deployer::Deployer.new(config, s3_client: s3_object_does_exist,
+lambda_client: lambda_client)
       uploader.update
     end
 
-    it 'does not update lambda code if code object does not exist' do
-      config.expects(:git_sha).returns('1234567890')
+    it "does not update lambda code if code object does not exist" do
+      config.expects(:git_sha).returns("1234567890")
       lambda_client.expects(:update_function_code).never
 
-      uploader = Deployer::Deployer.new(config, s3_client: s3_object_does_not_exist, lambda_client: lambda_client)
+      uploader = ServerlessTools::Deployer::Deployer.new(config, s3_client: s3_object_does_not_exist, lambda_client: lambda_client)
       uploader.update
     end
   end
